Refactor InspirationField to its own file
